requirements.txt
albumentations==0.4.3
numpy==1.21.6
opencv-python-headless==4.1.1.26
openslide-python==1.1.2
pandas==1.3.5
Pillow==8.4.0
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
tables==3.6.1
tifffile==2020.12.4
torch==1.7.1+cu110
tqdm==4.64.0
ttach==0.0.2
matplotlib==3.3.2
tensorboardX==1.9
